feat(usb): 支持STM32 OTG HS控制器及多控制器端点资源映射，完善接口/端点号参数化，补全远程唤醒与接口切换标准请求

+++ src/driver/usb/device/hid/hid.hpp
@@ -126,7 +126,7 @@ class HID : public DeviceClass
    * @param endpoint_pool 端点池 / Endpoint pool
    * @param start_itf_num 接口号起始 / Starting interface number
    */
-  void Init(EndpointPool& endpoint_pool, size_t start_itf_num) override
+  void Init(EndpointPool& endpoint_pool, uint8_t start_itf_num) override
   {
     inited_ = false;
     itf_num_ = start_itf_num;
